+++ src/Designer/frontend/packages/ux-editor/src/components/Settings/SettingsNavigation/ValidateNavigation/ValidateNavigationConfig.test.tsx
@@ -8,14 +8,13 @@ import {
 import userEvent, { type UserEvent } from '@testing-library/user-event';
 import { textMock } from '@studio/testing/mocks/i18nMock';
 
-const propertyLabel = 'Test Property';
+const undefinedLabel = textMock('ux_editor.settings.navigation_validation_button_rule_undefined');
 const allTasksCardLabel = textMock('ux_editor.settings.navigation_validation_all_tasks_card_label');
 
 describe('ValidateNavigationConfig', () => {
   it('renders add validation rule button when in view mode', () => {
     renderValidateNavigationConfig({ scope: Scope.AllTasks });
-
-    const addButton = screen.getByRole('button', { name: propertyLabel });
+    const addButton = screen.getByRole('button', { name: undefinedLabel });
     expect(addButton).toBeInTheDocument();
   });
 
@@ -37,9 +36,26 @@ describe('ValidateNavigationConfig', () => {
     ([action, label]) => {
       it(`should close config card when ${action} button is clicked`, async () => {
         const user = userEvent.setup();
-        renderValidateNavigationConfig({ scope: Scope.AllTasks });
-
+        renderValidateNavigationConfig({
+          scope: Scope.AllTasks,
+          config: {
+            types: [],
+            pageScope: { label: 'Page 1', value: 'page1' },
+          },
+        });
         await openCard(user);
+
+        if (action === 'save') {
+          const typeSelectorLabel = screen.getByText(
+            textMock('ux_editor.settings.navigation_validation_type_label'),
+          );
+          await user.click(typeSelectorLabel);
+          const typeOption = await screen.findByRole('option', {
+            name: textMock('ux_editor.component_properties.enum_Component'),
+          });
+          await user.click(typeOption);
+        }
+
         await user.click(screen.getByRole('button', { name: textMock(label) }));
         expect(screen.queryByText(allTasksCardLabel)).not.toBeInTheDocument();
       });
@@ -47,21 +63,17 @@ describe('ValidateNavigationConfig', () => {
   );
 });
 
-const openCard = async (user: UserEvent) => {
-  const addButton = screen.getByRole('button', { name: propertyLabel });
+const openCard = async (user: UserEvent, label?: string) => {
+  const addButton = screen.getByRole('button', { name: label });
   await user.click(addButton);
 };
 
 const renderValidateNavigationConfig = (props: Partial<ValidateNavigationConfigProps> = {}) => {
   const defaultProps: ValidateNavigationConfigProps = {
-    propertyLabel,
     scope: Scope.AllTasks,
     onSave: jest.fn(),
     onDelete: jest.fn(),
-    config: {
-      types: [],
-      pageScope: { value: '', label: '' },
-    },
+    config: undefined,
   };
   return render(<ValidateNavigationConfig {...defaultProps} {...props} />);
 };