+++ src/cli/CHANGELOG.md
@@ -9,4 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial implementation of studioctl CLI tool
+- Support `self update` and `self uninstall` for Linux, macOS
+
+### Fixed
+
+- PDF connectivity when running `env localtest` (#17959)
+- Handle partial "up" state in `env up` (#17959)
+
+## [0.1.0-preview.1] - 2026-02-25
+
+### Added
+
+- Initial implementation of studioctl CLI tool (#17841)